Additions and movement
I moved some of the stuff from deck to Player for players deck be able to add, remove, replace deck, and access the cards it has inside. The Players now have a chance to continue as long as they want and i will put the scoring additions and ACE rules after i make dinner. :+1: :+1: